fix: continue remaining osuTK to numerics migration
Agent-Logs-Url: https://github.com/winnerspiros/osu-framework/sessions/10551308-4e73-431f-84eb-ef77b3006f1c

+++ osu.Framework/Graphics/OpenGL/Shaders/GLShader.cs
@@ -196,7 +196,7 @@ namespace osu.Framework.Graphics.OpenGL.Shaders
 
                         case ResourceKind.StructuredBufferReadOnly:
                         case ResourceKind.StructuredBufferReadWrite:
-                            uniformBlocks[layout.Elements[0].Name] = GL4.GL.GetProgramResourceIndex(this, ProgramInterface.ShaderStorageBlock, layout.Elements[0].Name);
+                            uniformBlocks[layout.Elements[0].Name] = (int)GL4.GL.GetProgramResourceIndex(this, ProgramInterface.ShaderStorageBlock, layout.Elements[0].Name);
                             break;
                     }
                 }
